introduccion del size_t

+++ srcs/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: alejhern <alejhern@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 22:20:24 by alejhern          #+#    #+#             */
-/*   Updated: 2024/06/11 22:43:27 by alejhern         ###   ########.fr       */
+/*   Updated: 2024/06/15 23:28:47 by alejhern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	lenght;
+	size_t	lenght;
 
 	lenght = 0;
 	while (str[lenght])
